refactor(core): remove guardian compatibility aliases
