--- codex-rs/core/src/guardian/approval_request.rs
@@ -1,7 +1,0 @@
-pub(crate) use codex_guardian::format_guardian_action_pretty;
-#[cfg(test)]
-pub(crate) use codex_guardian::guardian_approval_request_to_json;
-pub(crate) use codex_guardian::guardian_assessment_action;
-pub(crate) use codex_guardian::guardian_request_target_item_id;
-pub(crate) use codex_guardian::guardian_request_turn_id;
-pub(crate) use codex_guardian::guardian_reviewed_action;
