+++ codex-rs/core/src/guardian/review.rs
@@ -38,11 +38,11 @@ use super::GuardianAssessment;
 use super::GuardianAssessmentOutcome;
 use super::GuardianRejection;
 use super::GuardianRejectionCircuitBreakerAction;
-use super::approval_request::guardian_assessment_action;
-use super::approval_request::guardian_request_target_item_id;
-use super::approval_request::guardian_request_turn_id;
-use super::approval_request::guardian_reviewed_action;
+use super::guardian_assessment_action;
+use super::guardian_request_target_item_id;
+use super::guardian_request_turn_id;
 use super::guardian_review_session_manager;
+use super::guardian_reviewed_action;
 use super::metrics::emit_guardian_review_metrics;
 use super::prompt::guardian_output_schema;
 use super::prompt::parse_guardian_assessment;
